requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=1.8.0
psycopg2-binary>=2.9.0
python-jose>=3.3.0
passlib>=1.7.0
python-multipart>=0.0.6
openai>=0.27.0
PyPDF2==3.0.1
python-docx==0.8.11